go test: move TestReset to civsession_test package
